Extract Extension from api module into extensions
